+++ AnotherNFCapp/.idea/misc.xml
@@ -5,6 +5,7 @@
       <map>
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.3489583333333333" />
         <entry key="app/src/main/res/layout/get_screen.xml" value="0.26197916666666665" />
+        <entry key="app/src/main/res/layout/login_screen.xml" value="0.26979166666666665" />
       </map>
     </option>
   </component>
