+++ AnotherNFCapp/app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">AnotherNFCapp</string>
+    <string name="password">Password</string>
 </resources>
